platform/api/aws: create network file
Move the SecurityGroup related methods to a new network file in
preparation for future changes to the AWS networking stack.
